SSR vs CSR and twitter and og stuff
efactor(projects): separate client and server components for metadata

BREAKING CHANGE: Projects page architecture has been restructured to properly
support OpenGraph/Twitter Card metadata generation.

- Split [slug]/page.tsx into server/client components
- Created new ProjectDetails.tsx client component
- Moved interactive logic (carousel, scroll) to client component
- Ensured metadata generation happens server-side
- Fixed social media preview images

This change resolves the Next.js error regarding 'use client' directive
conflicts with metadata generation, while maintaining all existing
functionality. The separation follows Next.js best practices for
handling server-side metadata with client-side interactivity.

Technical notes:
- ProjectDetails.tsx handles all client-side state and effects
- page.tsx focuses solely on server-side concerns (metadata, layout)
- No changes to underlying Convex data fetching
